Part 4
I created the Appointment class with the patient's name, phone number, and appointment time as well as polymorphism (the base class refers to different types of doctors), changed the base class to an abstract class, and forced the subclass to implement the getSpecialty() method. In the process, I also changed the getSpecialty() method in both subclasses.